+++ apps/frontend/src/data/schemas/user.schema.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod';
+
+export const UserSchema = z.object({
+  id: z.number(),
+  name: z.string(),
+  profile: z.string(),
+});
+
+export type TUser = z.infer<typeof UserSchema>;